chore: Purge all legacy standalone scripts and compatibility shims
